Iris
This project tackles the classic Iris dataset, aiming to classify iris flowers into one of three species: Setosa, Versicolor, or Virginica.
Also the project showcases classification in a straightforward machine learning context. KNN stood out with a 100% accuracy, but other models also performed admirably. It's an excellent reference for EDA, model training, and evaluation in the context of classification tasks.