Cars Price Prediction
The goal of this project is to predict the price of used cars based on various attributes and features. Factors like the brand's reputation, car features, horsepower, mileage, and more are considered to estimate the car's price accurately.
In summary, this code and analysis involve preprocessing a dataset of used car information, performing exploratory data analysis, one-hot encoding categorical variables, checking for multicollinearity, and training various regression models to predict car selling prices. The Gradient Boosting Regressor is identified as the best-performing model for this specific dataset.