Email Spam Detection
The dataset used for this project is the "SMS Spam Collection," which consists of SMS messages tagged as either "ham" (legitimate) or "spam." The aim is to build an email spam detector using machine learning.
In summary, firstly we do analysis to preprocess text data, train machine learning models to detect spam emails, and visualize common words in spam messages. The Support Vector Machine (SVM) model appears to perform the best, with high precision and recall for detecting spam.