Sales Predictions
This dataset is related to sales prediction based on advertising expenditure in different media channels (TV, Radio, and Newspaper). I have explored, visualized, and built several regression models to predict sales. I have made predictions using a sample of 5 data points from the dataset for each of the models. The predictions were compared with the original sales values. 
In result , we get that the gradient boosting regressor and random forest regressor are the top-performing models, with high R2 scores and low mean squared errors. These models provide the best predictions for the given dataset.